# webapps/swagger.yaml
swagger: '2.0'
info:
  description: "This website for testing API of Test server project ISS. If any question can read more detail on http://swagger.io or on irc.freenode.net, #swagger."
  version: "1.0.0"
  title: "Swagger ISS"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "issadmin@celestica.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
consumes:
  - application/json
produces:
  - application/json
basePath: "/api"
...
paths:
...
  /checkout:
    post:
      tags:
      - "Testing"
      summary: "Request to checkout test"
      description: ""
      operationId: "checkout"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Serial number is using delivery out test user must be provide via website."
        required: true
        schema:
          $ref: "#/definitions/Checkout"
      responses:
        200:
          description: "successful delivery out testing"
        401:
          description: "Input is not found"
        405:
          description: "Input is invalid format"
      x-swagger-router-controller: "controllers.test_controller"
...
  /statuses:
    get: 
      tags: [Statuses]
      operationId: controllers.status_controller.get_statuses
      summary: Get all statuses
      parameters: 
        - name: limit
          in: query
          type: integer
          minimum: 0
          default: 100
      responses:
        200:
          description: Return status
          schema:
            type: array
            items:
              $ref: '#/definitions/Status'
    put:  
      tags: [Statuses]
      operationId: controllers.status_controller.put_status
      summary: Create or update a status
      parameters: 
        - name: status
          in: body
          schema:
            $ref: '#/definitions/Status'
      responses:
        200:
          description: status updated
        201:
          description: New status created
  /statuses/{status_id}:
    get:  
      tags: [Statuses]
      operationId: controllers.status_controller.get_status
      summary: Get a single status
      parameters:
        - $ref: '#/parameters/status_id'
      responses:
        200:
          description: Return status
          schema:
            $ref: '#/definitions/Status'
        404:
          description: Status does not exist
    delete:  
      tags: [Statuses]
      operationId: controllers.status_controller.delete_status
      summary: Remove a status
      parameters:
        - $ref: '#/parameters/status_id'
      responses:
        204:
          description: Status was deleted
        404:
          description: Status does not exist
  /statuses/slot/{location}:
      get: 
        tags: [Statuses]
        operationId: controllers.status_controller.get_status_by_slot
        summary: Get all statuses by slot
        parameters: 
          - $ref: '#/parameters/location' 
        responses:
          200:
            description: Return status
            schema:
              type: array
              items:
                $ref: '#/definitions/Status'
  /statuses/update_slot/{location}:
      get: 
        tags: [Statuses]
        operationId: controllers.status_controller.update_data
        summary: Get all statuses by slot
        parameters: 
          - $ref: '#/parameters/location' 
        responses:
          200:
            description: Return status
            schema:
              type: array
              items:
                $ref: '#/definitions/Status'
...
parameters:
...
  status_id:
    name: status_id
    description: Status's Unique identifier
    in: path
    type: integer
    required: true
  location:
    name: location
    description: slot number of test server
    in: path
    type: integer
    required: true
definitions:
...
  Checkout:
    type: "object"
    required:
    - "serial_number"
    - "test_mode"
    - "operation_id"
    - "slot_location"
    - "test_station"
    - "odc_type"
    - "uut_log_dir"
    properties:
      serial_number:
        type: "string"
      test_mode:
        type: "string"
      operation_id:
        type: "string"
      slot_location:
        type: "string"
      test_station:
        type: "string"
      odc_type:
        type: "string"
      uut_log_dir:
        type: "string"
    example:
      serial_number: "CF2SZ183400021"
      test_mode: "Production"
      operation_id: "th00000000"
      slot_location: "ChassisSN_slotNo"
      test_station: "Pre_FCT"
      odc_type: "verify"
      uut_log_dir: "uut_log_dir_simple"
...
  Status:
    type: object
    required:
      - status
    properties: 
      status:
        type: string
        description: Status's name
        example: "PASS"
        minLength: 1
        maxLength: 100
      message:
        type: string
        description: Status's message
        example: "print................"
        minLength: 0 
        maxLength: 1000
      test_location:
        type: string
        description: Location of Test
        example: "PF01_01"  
...